Added import/export/delete_all

+++ app/Http/Controllers/AttendanceController.php
@@ -2,11 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\AttendancesExport;
+use App\Imports\AttendancesImport;
+use App\Models\Role;
 use App\Models\Room;
 use App\Models\Course;
 use App\Models\Student;
 use App\Models\Attendance;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\ValidationException;
+use Maatwebsite\Excel\Facades\Excel;
 
 class AttendanceController extends Controller
 {
@@ -15,8 +21,12 @@ class AttendanceController extends Controller
      */
     public function index()
     {
+        if (Gate::denies('view', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
+        $roles = Role::all();
         $attendances = Attendance::with('student', 'room')->paginate(5);
-        return view('admin.attendance.index', compact('attendances'));
+        return view('admin.attendance.index', compact('attendances','roles'));
     }
 
     /**
@@ -24,6 +34,9 @@ class AttendanceController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
         $students = Student::all();
         $courses = Course::all();
         $rooms = Room::all();
@@ -35,6 +48,9 @@ class AttendanceController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('create', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
         $request->validate([
             'student_name' => 'required',
             'course_name' => 'required',
@@ -66,6 +82,9 @@ class AttendanceController extends Controller
      */
     public function edit(string $id)
     {
+        if (Gate::denies('update', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
         $attendance = Attendance::find($id);
         return view('admin.attendance.edit', compact('attendance'));
     }
@@ -75,6 +94,9 @@ class AttendanceController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Gate::denies('update', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
         $request->validate([
             'student_name' => 'required',
             'course_name' => 'required',
@@ -99,8 +121,70 @@ class AttendanceController extends Controller
      */
     public function destroy(string $id)
     {
+        if (Gate::denies('delete', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
         $attendance = Attendance::find($id);
         $attendance->delete();
         return redirect()->route('attendances.index')->with('success', 'One row deleted!');
+    }
+
+    public function destroyall()
+    {
+        if (Gate::denies('delete', Attendance::class)) {
+            return redirect()->route('admin.dashboard')->with('error', 'No permission.');
+        }
+        $attendances = Attendance::all();
+        if ($attendances->isEmpty()) {
+            return redirect()->route('attendances.index')->with('error', 'No attendances to delete.');
+        }
+        foreach ($attendances as $attendance) {
+            $attendance->delete();
+        }
+        return redirect()->route('attendances.index')->with('success','All attendances deleted successfully.');
+    }
+
+    public function search(Request $request){
+        $searchData = $request->search_data;
+        if($searchData == ""){
+            return redirect()->route('attendances.index');
+        } else {
+            $attendances = Attendance::whereHas('student', function($students) use ($searchData){
+                $students->where('name','LIKE','%'.$searchData.'%');
+            })
+            ->orwhereHas('course', function($query) use ($searchData){
+                $query->where('name','LIKE','%'.$searchData.'%');
+            })
+            ->orwhereHas('room', function($rooms) use ($searchData){
+                $rooms->where('name','LIKE','%'.$searchData.'%');
+            })
+            ->orWhere('attendance_date', 'LIKE', '%'.$searchData.'%')
+            ->orWhere('attendance_status', 'LIKE', '%'.$searchData.'%')
+            ->paginate(5);
+            $roles = Role::all();
+            return view('admin.attendance.index', compact('attendances', 'roles'));
+        }
+    }
+
+    public function export(){
+        return Excel::download(new AttendancesExport(),'attendances.xlsx');
+    }
+
+    public function import(Request $request){
+        $request->validate(['attendances'=>'required|file|mimes:xlsx,xls,csv']);
+        $file = $request->file('attendances');
+        $originalName = $file->getClientOriginalName();
+
+        if (!str_contains(strtolower($originalName), 'attendances')) {
+            return redirect()->back()->with('error', 'Excel file name must contain the word "attendances"');
+        }
+        try {
+            Excel::import(new AttendancesImport(), $request->file('attendances'));
+            return redirect()->route('attendances.index')->with('success', 'Attendances imported successfully!');
+        } catch (ValidationException $e) {
+            return back()->with('error', $e->getMessage());
+        } catch (\Exception $e) {
+            return back()->with('error', 'Failed to import. Please check your Excel file format.');
+        }
     }
 }
