+++ app/Http/Controllers/TeacherCoursesController.php
@@ -2,10 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\TeacherCoursesExport;
+use App\Imports\TeacherCoursesImport;
 use App\Models\Course;
+use App\Models\Role;
 use App\Models\Teacher;
 use Illuminate\Http\Request;
 use App\Models\TeacherCourse;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\ValidationException;
+use Maatwebsite\Excel\Facades\Excel;
 
 class TeacherCoursesController extends Controller
 {
@@ -14,8 +20,12 @@ class TeacherCoursesController extends Controller
      */
     public function index()
     {
+        if (Gate::denies('view', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
+        $roles = Role::all();
         $teacher_courses = TeacherCourse::with(['teacher', 'course'])->paginate(5);
-        return view('admin.teacher_course.index', compact('teacher_courses'));
+        return view('admin.teacher_course.index', compact('teacher_courses', 'roles'));
     }
 
     /**
@@ -23,6 +33,9 @@ class TeacherCoursesController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teachers = Teacher::all();
         $courses = Course::all();
         return view('admin.teacher_course.create', compact('teachers', 'courses'));
@@ -33,6 +46,9 @@ class TeacherCoursesController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('create', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'teacher_name' => 'required',
             'course_name' => 'required',
@@ -58,6 +74,9 @@ class TeacherCoursesController extends Controller
      */
     public function edit(string $id)
     {
+        if (Gate::denies('update', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teachers = Teacher::all();
         $courses = Course::all();
         $teacher_course = TeacherCourse::find($id);
@@ -69,6 +88,9 @@ class TeacherCoursesController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Gate::denies('update', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'teacher_name' => 'required|string|max:30',
             'course_name' => 'required|string|max:50',
@@ -86,8 +108,65 @@ class TeacherCoursesController extends Controller
      */
     public function destroy(string $id)
     {
+        if (Gate::denies('delete', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teacher_course = TeacherCourse::find($id);
         $teacher_course->delete();
         return redirect()->route('teachercourses.index')->with('success', 'One row Deleted!');
+    }
+
+    public function destroyall()
+    {
+        if (Gate::denies('delete', TeacherCourse::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
+        $teacher_courses = TeacherCourse::all();
+        if ($teacher_courses->isEmpty()) {
+            return redirect()->route('teachercourses.index')->with('error', 'No teacher courses to delete.');
+        }
+        foreach ($teacher_courses as $teacher_course) {
+            $teacher_course->delete();
+        }
+        return redirect()->route('teachercourses.index')->with('success','All teacher courses deleted successfully.');
+    }
+
+    public function search(Request $request){
+        $searchData = $request->search_data;
+        if($searchData == ""){
+            return redirect()->route('teachercourses.index');
+        } else {
+            $teachercourses = TeacherCourse::whereHas('course', function($courses) use ($searchData){
+                $courses->where('name','LIKE','%'.$searchData.'%');
+            })
+            ->orWhereHas('teacher', function($teachers) use ($searchData){
+                $teachers->where('name','LIKE','%'.$searchData.'%');
+            })
+            ->paginate(5);
+            $roles = Role::all();
+            return view('admin.teacher_course.index', compact('teachercourses', 'roles'));
+        }
+    }
+
+    public function export(){
+        return Excel::download(new TeacherCoursesExport(),'teachercourses.xlsx');
+    }
+
+    public function import(Request $request){
+        $request->validate(['teachercourses'=>'required|file|mimes:xlsx,xls,csv']);
+        $file = $request->file('teachercourses');
+        $originalName = $file->getClientOriginalName();
+
+        if (!str_contains(strtolower($originalName), 'teachercourses')) {
+            return redirect()->back()->with('error', 'Excel file name must contain the word "teachercourses"');
+        }
+        try {
+            Excel::import(new TeacherCoursesImport(), $request->file('teachercourses'));
+            return redirect()->route('teachercourses.index')->with('success', 'Teacher courses imported successfully!');
+        } catch (ValidationException $e) {
+            return back()->with('error', $e->getMessage());
+        } catch (\Exception $e) {
+            return back()->with('error', 'Failed to import. Please check your Excel file format.');
+        }
     }
 }
